backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
groq==0.11.0
python-dateutil==2.8.2
Werkzeug==3.0.1
feedparser==6.0.10
